Changed the projected savings data type to a float

+++ python_introduction/finance_calculator.py
@@ -5,7 +5,7 @@ total_monthly_expenses = int(input("Enter your total monthly expenses: "))
 monthly_savings = monthly_income - total_monthly_expenses
 
 # Calculating the projected savings after one year, incorporating the interest
-projected_savings = monthly_savings * 12 + (monthly_savings * 12 * 0.05)
+projected_savings = float(monthly_savings * 12 + (monthly_savings * 12 * 0.05))
 
 # Printing the user's monthly savings and projected annual savings after including interest
 print("Your monthly savings are ${}".format(monthly_savings))
